# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KATALA - Katalog & Reservasi Wisata Lampung API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /destinations:
    get:
      operationId: listDestinations
      tags: [destinations]
      summary: List all destinations
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of destinations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationListResponse"
    post:
      operationId: createDestination
      tags: [admin]
      summary: Create a destination (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
  /destinations/{id}:
    get:
      operationId: getDestination
      tags: [destinations]
      summary: Get a destination by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Destination detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
        "404":
          description: Not found
    patch:
      operationId: updateDestination
      tags: [admin]
      summary: Update a destination (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
    delete:
      operationId: deleteDestination
      tags: [admin]
      summary: Delete a destination (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ─── AUTH ─────────────────────────────────────────────
...
    Destination:
      type: object
      required: [id, name, description, location, latitude, longitude, price, dailyQuota, category, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        location:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        price:
          type: number
        dailyQuota:
          type: integer
        category:
          type: string
        imageUrl:
          type: string
        rating:
          type: ["number", "null"]
        featured:
          type: boolean
        createdAt:
          type: string
    DestinationInput:
      type: object
      required: [name, description, location, latitude, longitude, price, dailyQuota, category, imageUrl]
      properties:
        name:
          type: string
        description:
          type: string
        location:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        price:
          type: number
        dailyQuota:
          type: integer
        category:
          type: string
        imageUrl:
          type: string
        rating:
          type: ["number", "null"]
        featured:
          type: boolean
    DestinationUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        location:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        price:
          type: number
        dailyQuota:
          type: integer
        category:
          type: string
        imageUrl:
          type: string
        rating:
          type: ["number", "null"]
        featured:
          type: boolean
    DestinationListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Destination"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── RESERVATIONS ──────────────────────────────────────
...