# lib/api-zod/src/generated/types/destinationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KATALA - Katalog & Reservasi Wisata Lampung API
 * OpenAPI spec version: 0.1.0
 */

export interface DestinationInput {
  name: string;
  description: string;
  location: string;
  latitude: number;
  longitude: number;
  price: number;
  dailyQuota: number;
  category: string;
  imageUrl: string;
  /** @nullable */
  rating?: number | null;
  featured?: boolean;
}