# lib/api-zod/src/generated/types/destinationListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KATALA - Katalog & Reservasi Wisata Lampung API
 * OpenAPI spec version: 0.1.0
 */
import type { Destination } from "./destination";

export interface DestinationListResponse {
  data: Destination[];
  total: number;
  page: number;
  limit: number;
}